+ py: don't crash when user stops debugger
often we will get what we need to know and then exit the debugger.  this
will cause an exception in the program and crash it.  we'll just catch
it and print what happened, because this really isn't a crash condition,
it's a known thing that we requested ourselves while in the debugger
(with ctrl-d, etc)

hopefully the import of exception doesn't cause a measurable speed
penalty.  we could measure it, but we can't remove it anyways because we
want it there.  no way to conditionally import it because it surrounds
main(), we'd have two different main() calls and that would be strange
just to avoid the import.  doubt this matters