+ py: dprintvar: util func to pretty print var and its name
avoids:

    dprint(f"var: {var}")

but the replacement is really just as long:

    dprintvar('var', vars())

but it's shorter for longer variable names.  but it also pretty-prints
the variable using pprint.pprint().

have to think about it, maybe this is dumb and it will be removed later,
but was debugging a bunch of stuff that had a lot of duplicated strings
for variable name as a string and variable name in the namespace, and
thought it was silly to duplicate all that every call