+ py: invargs: addopts: new argparse wrapper for option argument
there was no way to --option=foo with existing wrappers.  this adds one
and makes a common function addopt that both addflag and addopts is
implemented in terms of.  it only takes one argument though, so maybe it
shouldn't be named 'addopts' but this is in keeping with the same thing
we did in last commit with 'addargs' wrapping the common 'addarg'.
maybe we'll change the naming later...