+ py: process_args: unesc: revert escaping in argparse escaped strings
sometimes we want raw strings available to process, exactly as specified
by the user, without escaping anything, which is what python does in
argparse.ArgumentParser.parse_args().  example, when the argument is
supposed to be itself a python string literal (we do this in 'field' for
fixed-string patterns).  not typically useful, but it should be
available in the template because in some programs this is useful...