+ py: process_args: hasopt: new function to see if opt was provided
argparse uses a SimpleNamespace and if args it knows about are not
provided by the user, it will set them to None.  sometimes we have a
whole class of options, for any of which we will want to change the
program behavior.  so we provide a convenience function hasopt() that
can take a list of options to check for and if any are set it will
return true