/ py: dmain: finally: exception backstop must be in global context
in practice, we have found that if the exception backstop (ie, catcher
of last resort) is not in global context, but instead is inside our root
driver function, the last exception gets eaten and we never get a
printout (although the program does crash).  this is presumably because
stdio gets closed before built-in exception handling does its prints,
which occurs in global context as the program exits, and top-level error
handling happens.

so, we have to catch it ourselves and print the exception.  so this
last-resort nested try/finally block belongs in dmain rather than main.

this took a while to figure out and get to work right in different
scenarios.