/ py: dmain: select: rather than timeout 0, check if stdin isatty(3)
when checking for input, we had been testing whether stdin was ready to
read immediately.  sometimes this is not the case, such as when the
program writing to us in a pipeline does some processing first, we may
get to the select(2) before any input is ready.  then, since we had set
timeout zero, the call returns immediately and we are forced to conclude
that no stdin is available -- an error if it was expected -- but that is
not necessarily correct.  this leads to naive case like "echo foo |
program" working correctly but "(sleep 1; echo foo) | field" will fail.
quite a few programs that would run in a pipeline do not have output
available to write yet by the time we are doing select().

instead, a better solution is to use isatty(), and test if stdin is a
terminal.  if not, we must be in a pipeline or be receiving by redirect.
this is how typical unix filter programs do it.