/ py: dmain: lastresort: exit via bomb() so there's a bad status
because we have a last-resort exception handler (which merely prints the
last traceback), we'd get a zero exit status, unless we explicitly exit
with a status ourselves.  if the program has crashed with unhandled
exception, it should definitely have a bad exit.

this was causing us problems when using the block in a program that
relied on exit status

while we're at it we print an "aborting..." message, at least to show
that it was caught, and how we're proceeding