/ invargs: dmain: lastresort: when doing close ignore stderr flush err
without this, crash is produced in certain scenarios, something in
stderr flush had an error or something and tried to print to stdout
rather than stderr, it was some internal python thing that should never
have happened and obvious internal python bug.  tried switching the
order of close to do stderr first then stdout, but this had a similar
issue itself so there was never a proper sequence for it that worked in
all cases.  it had something to do with pipe closing unexpecedtly so
it's already an error condition and we'll lose data anyhow

too bad don't have a stack trace anymore, computer rebooted since in a
bit of an unexpected way and was never captured.  one day we might
remove the line in a script and try to reproduce it but for now we'll
just keep it in place, should not cause a problem because stdout is
already flushed and this only happens when the output is already hosed