+ py: add a __devskel__ top level so we can keep track in users
we often update the devskel and have to kind of hand-compare to see the
level we're at.  what we'll do is add a top-level __devskel__ key and
maintain it with the unique sha of the latest git version, and embed
this in all our scripts that use it.  in theory devskel won't actually
change much, it's fairly stabilized by now, so this shouldn't cause too
much of a hassle, and it will also allow us to programmatically check
users (in combination with eg our 'srcdirs' command)