/ py: devskel: make the dunder var a string corresponding to tag
well intentioned but, silly: we can't use a commit hash, because the
actual commit is not known prior to the commit itself, so there's no way
to actually embed it in the file.  however we *can* predict the tag,
since we can create it ourselves after the commit, and know our
anticipated tag name before we make the commit, so we can use that
instead.