- py: imports: traceback: move import to the site since it's catch-all
this import is not generally useful, leave it to the specific case that
we hit the last-resort exception.  this is only really an issue if we're
out of memory and if that happens (1) we're already going to crash, and
(2) we're ok with whatever default unhandled exception scenario this
would cause, like maybe printing that we're out of memory.  it should
really never happen.  so we'll avoid polluting the import block by
keeping this in its own dmain section where it's used, just like we did
with bdb/pp import.  this is arguably not pythonic but that's ok, it's
more readable for us