+ py: exe: add alternate impl, uses subprocess.run(), handles exceptions
use this for more complicated scenarios, it allows greater control,
allowing kwarg defaults override, and handles its own errors.  it also
has some comments to return things other that stdout if that's what the
user wants.  we'll intend this to be used for more sophisticated cases.
probably we need to use Popen directly for the most sophisticated ones
but run() should suffice for nearly everything