- sh: usagex: rename from usage_exit, keeps with our usual pattern
and give it a bad exit code.  since usage() itself is available, we can
use that if we don't want to have a bad exit.  maybe we'll reverse this
decision later and have it pass through $? though.  but then it requires
all the users to do a 'false' first whereas 'true' is the exception