+ sh: usage: add alternate implementation that uses one sed command
maybe we'll improve on this in the future.  it has one issue, it will
always have an extra blank line at the bottom.  for smaller scripts or
if we don't like feeling the kind of hack already implemented, we'll
have this hack instead and think about replacing it as the main one,
especially if we can figure out a way to remove the trailing blank line