^ sh: exewrap: remove the same-name binary-wrapper code, does not belong
this belong as a separate script maybe but shouldn't be in the standard
template.  we have only used it so far for tmux and taskwarrior, and it
isn't common.  possibly we should have it in a separate file in devskel
but we'll think about it.  for now it's fine to remove it because we can
always look at the tmux one for the canonical example.